requirements.txt
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.1
Pillow==10.0.1
Pillow==10.2.0
streamlit==1.31.1
